Add function to play a round of the game
Added the function playRound() to play a round of Rock Paper Scissors.
Also created an additional function which capitalizes only the first
letter of a word. This was to make every single input case insensitive,
such that, for example no matter if the user entered
"Rock, ROCK, rock, RoCk", the value would always be "Rock".